pyproject.toml
[tool.poetry]
name = "drf-api-permissions"
version = "0.1.0"
description = ""
authors = ["Griffin <60614863+h-griffin@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.7"
djangorestframework = "^3.11.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
